Fix: Client compilation errors for Minecraft 1.21.1
Bugfixes:
- FuelHudOverlay: height von float zu double geändert (getY() returns double)
- ShipEntityRenderer: playSound zu playLocalSound geändert (korrekte Signatur)
- ModSounds: Holder<SoundEvent> zu SoundEvent vereinfacht (Fabric auto-register)
- ShipEntity: getY() Cast zu float für calculateHeightPenalty

Changes:
- 4 Dateien geändert
- 21 Zeilen hinzugefügt, 34 Zeilen entfernt
- Client kompiliert jetzt erfolgreich

Status: Client läuft und ist spielbar ✅

+++ sharkengine/src/client/java/dev/sharkengine/client/render/FuelHudOverlay.java
@@ -63,7 +63,7 @@ public final class FuelHudOverlay {
         int fuel = ship.getFuelLevel();
         int blocks = ship.getBlockCount();
         float speed = ship.getCurrentSpeed();
-        float height = player.getY();
+        double height = player.getY();
         WeightCategory weight = ship.getWeightCategory();
         
         // Render HUD background (semi-transparent black)
@@ -140,7 +140,7 @@ public final class FuelHudOverlay {
      * @param speed Current speed
      * @param height Current height (Y position)
      */
-    private static void renderStats(GuiGraphics g, Minecraft mc, int x, int y, int blocks, float speed, float height) {
+    private static void renderStats(GuiGraphics g, Minecraft mc, int x, int y, int blocks, float speed, double height) {
         // Height
         String heightText = String.format("Höhe: Y=%.0f", height);
         g.drawString(mc.font, heightText, x, y, TEXT_COLOR);
